+++ packages/editor/web/src/RichContentEditor/InnerRCE.jsx
@@ -8,7 +8,6 @@ import 'wix-rich-content-common/dist/statics/styles/draftDefault.rtlignore.scss'
 import { LINK_PREVIEW_TYPE, TABLE_TYPE } from 'wix-rich-content-common';
 import { cloneDeep } from 'lodash';
 import { isCursorAtStartOfContent, selectAllContent } from 'wix-rich-content-editor-common';
-import ClickOutside from 'react-click-outsider';
 
 class InnerRCE extends PureComponent {
   constructor(props) {
@@ -16,13 +15,14 @@ class InnerRCE extends PureComponent {
     const { innerRCERenderedIn, config } = props;
     this.config = this.cleanConfig(cloneDeep(config));
     this.plugins = config[innerRCERenderedIn].innerRCEPlugins;
-    this.state = {
-      showToolbars: false,
-    };
   }
 
   componentDidUpdate(prevProps) {
-    if (prevProps.editing === false && prevProps.editing !== this.props.editing) {
+    if (
+      this.props.innerRCERenderedIn === 'table' &&
+      prevProps.editing === false &&
+      prevProps.editing !== this.props.editing
+    ) {
       this.handleAtomicPluginsBorders(true);
     }
   }
@@ -51,26 +51,6 @@ class InnerRCE extends PureComponent {
   onChange = editorState => {
     this.props.onChange(editorState);
     this.editorHeight = this.editorWrapper.offsetHeight;
-  };
-
-  onFocus = e => {
-    e.stopPropagation();
-    this.props.setEditorToolbars(this.ref);
-    this.props.setInPluginEditingMode(true);
-    this.setState({ showToolbars: true });
-  };
-
-  onClickOutside = e => {
-    if (
-      this.state.showToolbars &&
-      this.editorWrapper &&
-      e.target &&
-      !e.target.closest('[data-id=rich-content-editor-modal]') &&
-      !e.target.closest('[class=ReactModalPortal]') &&
-      !this.editorWrapper.contains(e.target)
-    ) {
-      this.setState({ showToolbars: false });
-    }
   };
 
   getToolbars = () => {
@@ -112,13 +92,12 @@ class InnerRCE extends PureComponent {
     e.stopPropagation();
     this.ref && this.props.setEditorToolbars(this.ref);
     this.props.setInPluginEditingMode(true);
-    this.setState({ showToolbars: true });
+    this.props.onFocus && this.props.onFocus();
   };
 
   handleAtomicPluginsBorders = enterEditing => {
     const { editing = true } = this.props;
-    const { showToolbars } = this.state;
-    const hideBorder = !showToolbars || !editing;
+    const hideBorder = !editing;
     if (this.editorWrapper) {
       const atomicBlocksNodeList = this.editorWrapper.querySelectorAll('[data-focus]');
       const atomicBlocks = Array.apply(null, atomicBlocksNodeList);
@@ -144,40 +123,37 @@ class InnerRCE extends PureComponent {
       tablePluginMenu,
       ...rest
     } = this.props;
-    const { showToolbars } = this.state;
     this.handleAtomicPluginsBorders();
     if (innerRCERenderedIn === TABLE_TYPE && isMobile) {
       toolbarsToIgnore.push('SideToolbar');
     }
     return (
-      <ClickOutside onClickOutside={this.onClickOutside}>
-        <div
-          data-id="inner-rce"
-          onFocus={this.onFocus}
-          className={classNames(styles.editor, theme.editor, 'inner-rce')}
-          ref={this.setEditorWrapper}
-        >
-          <RichContentEditor
-            {...rest} // {...rest} need to be before editorState, onChange, plugins
-            ref={this.setRef}
-            editorState={editorState}
-            onChange={this.onChange}
-            plugins={this.plugins}
-            config={this.config}
-            isMobile={isMobile}
-            toolbarsToIgnore={['FooterToolbar', ...toolbarsToIgnore]}
-            showToolbars={editing && showToolbars}
-            isInnerRCE
-            innerRCERenderedIn={innerRCERenderedIn}
-            editorKey="inner-rce"
-            readOnly={readOnly}
-            onBackspace={this.onBackspaceAtBeginningOfContent}
-            direction={direction}
-            tablePluginMenu={tablePluginMenu}
-            {...additionalProps}
-          />
-        </div>
-      </ClickOutside>
+      <div
+        data-id="inner-rce"
+        onFocus={this.onFocus}
+        className={classNames(styles.editor, theme.editor, 'inner-rce')}
+        ref={this.setEditorWrapper}
+      >
+        <RichContentEditor
+          {...rest} // {...rest} need to be before editorState, onChange, plugins
+          ref={this.setRef}
+          editorState={editorState}
+          onChange={this.onChange}
+          plugins={this.plugins}
+          config={this.config}
+          isMobile={isMobile}
+          toolbarsToIgnore={['FooterToolbar', ...toolbarsToIgnore]}
+          showToolbars={editing}
+          isInnerRCE
+          innerRCERenderedIn={innerRCERenderedIn}
+          editorKey="inner-rce"
+          readOnly={readOnly}
+          onBackspace={this.onBackspaceAtBeginningOfContent}
+          direction={direction}
+          tablePluginMenu={tablePluginMenu}
+          {...additionalProps}
+        />
+      </div>
     );
   }
 }
@@ -200,6 +176,7 @@ InnerRCE.propTypes = {
   toolbarsToIgnore: PropTypes.array,
   editing: PropTypes.bool,
   tablePluginMenu: PropTypes.bool,
+  onFocus: PropTypes.func,
 };
 
 export default InnerRCE;