+++ packages/plugin-table/web/src/table-component.jsx
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import TableViewer from './table-viewer';
 import styles from '../statics/styles/table-component.scss';
 import Table from './domain/table';
-import { getRange } from './domain/tableDataUtil';
+import { getRange, getRowsRange, getColsRange } from './domain/tableDataUtil';
 import { createEmptyCellEditor, handleCellClipboardEvent } from './tableUtil';
 import { AddNewSection, Rows } from './components';
 import TableToolbar from './TableToolbar/TableToolbar';
@@ -25,7 +25,7 @@ class TableComponent extends React.Component {
     };
     this.innerRceAdditionalProps = { placeholder: '' };
     this.innerEditorsRefs = {};
-    this.table = new Table(props.componentData, this.updateComponentData1);
+    this.table = new Table(props.componentData, this.updateComponentData);
     this.tableRef = createRef();
     this.dragPreview = createRef();
     this.rowDragProps = {
@@ -85,6 +85,16 @@ class TableComponent extends React.Component {
   getCellEditorRef = (i, j) => this.innerEditorsRefs[`${i}-${j}`];
 
   getCellToolbarProps = (i, j) => this.getCellEditorRef(i, j)?.getToolbarProps(TOOLBARS.FORMATTING);
+
+  distributeRows = selected => {
+    this.table.distributeRows(this.innerEditorsRefs, getRowsRange(selected));
+    this.setSelected();
+  };
+
+  distributeColumns = selected => {
+    this.table.distributeColumns(getColsRange(selected));
+    this.setSelected();
+  };
 
   isCellEmpty = (i, j) =>
     this.table
@@ -162,32 +172,41 @@ class TableComponent extends React.Component {
     return firstSelectedCellRef;
   };
 
+  setCellContent = (content, row, col) => {
+    const rows = this.table.getRows();
+    this.table.setCellContent(rows, content, row, col);
+    this.table.setNewRows(rows);
+  };
+
   handleTableClipboardEvent = e => {
-    const { selected, isEditingActive, copiedCellsRange, isAllCellsSelected } = this.state;
-    if (isPluginFocused(this.props.block, this.props.selection) && selected && !isEditingActive) {
+    const { selected, isEditingActive } = this.state;
+    const isColorPickerModalOpen = e.target.closest('[data-id=color-picker-modal]');
+    if (
+      isPluginFocused(this.props.block, this.props.selection) &&
+      selected &&
+      !isEditingActive &&
+      !isColorPickerModalOpen
+    ) {
       const preventEvent = () => {
         e.stopPropagation();
         e.preventDefault();
       };
-      const isColorPickerModalOpen = e.target.closest('[data-id=color-picker-modal]');
-      if (e.key === 'Backspace' && !isAllCellsSelected && !isColorPickerModalOpen) {
-        preventEvent();
-        this.table.clearRange(getRange(selected));
-      } else if (e.key === 'a' && (e.ctrlKey || e.metaKey)) {
+      if (e.key === 'a' && (e.ctrlKey || e.metaKey)) {
         preventEvent();
         this.setAllCellsSelected();
-      } else if (copiedCellsRange && e.key === 'v' && (e.ctrlKey || e.metaKey)) {
-        preventEvent();
-        this.table.pasteCells(copiedCellsRange, selected.start.i, selected.start.j);
-        this.setSelectionOnPastedCells();
       }
     }
   };
 
-  setSelectionOnPastedCells = () => {
-    const { selected, copiedCellsRange } = this.state;
-    const { i: startI, j: startJ } = copiedCellsRange[0];
-    const { i: endI, j: endJ } = copiedCellsRange[copiedCellsRange.length - 1];
+  onPaste = (copiedCells, start) => {
+    this.table.pasteCells(copiedCells, start.i, start.j);
+    this.setSelectionOnPastedCells(copiedCells);
+  };
+
+  setSelectionOnPastedCells = copiedCells => {
+    const { selected } = this.state;
+    const { i: startI, j: startJ } = copiedCells[0];
+    const { i: endI, j: endJ } = copiedCells[copiedCells.length - 1];
     const copiedRowsNum = endI - startI + 1;
     const copiedColsNum = endJ - startJ + 1;
     this.setSelected({
@@ -214,18 +233,19 @@ class TableComponent extends React.Component {
     !this.isAllCellsSelected(this.state.selected) && this.setState({ isAllCellsSelected: false });
   };
 
-  updateComponentData1 = data =>
+  updateComponentData = data =>
     this.props.store.set('componentData', { ...data }, this.props.block.getKey());
 
   onResizeCol = columnsRefs =>
     this.table.setColWidthAfterResize(columnsRefs, this.tableRef.current.offsetWidth);
 
   onResizeRow = (i, height) =>
-    this.table.setRowHeight(getRange(this.table.getRowsSelection({ start: i, end: i })), height);
+    this.table.setRowHeight(
+      getRowsRange(this.table.getRowsSelection({ start: i, end: i })),
+      height
+    );
 
   setToolbarRef = ref => (this.toolbarRef = ref);
-
-  handleCopy = ({ end, start }) => this.setState({ copiedCellsRange: getRange({ start, end }) });
 
   highlightResizer = (i, isCol) => {
     if (isCol && this.state.highlightColResizer !== i) {
@@ -390,7 +410,6 @@ class TableComponent extends React.Component {
         className={classNames(styles.tableEditorContainer, 'has-custom-focus', {
           [styles.editMode]: isEditMode,
           [styles.viewMode]: !isEditMode,
-          [styles.disableSelection]: !isEditingActive,
         })}
         data-hook="TableComponent"
         onFocus={this.onFocus}
@@ -407,7 +426,7 @@ class TableComponent extends React.Component {
             deleteColumn={this.deleteColumn}
             deleteRow={this.deleteRow}
             isEditingActive={isEditingActive}
-            tableWidth={this.tableRef.current?.offsetWidth}
+            tableWidth={this.tableRef.current?.parentElement?.offsetWidth}
             getFirstCellRef={this.getFirstCellRef}
             t={t}
             theme={theme}
@@ -418,6 +437,8 @@ class TableComponent extends React.Component {
             deleteBlock={blockProps.deleteBlock}
             isAllCellsSelected={this.isAllCellsSelected(selected)}
             merge={this.merge}
+            distributeRows={this.distributeRows}
+            distributeColumns={this.distributeColumns}
           />
         )}
         {!isMobile && (
@@ -454,7 +475,6 @@ class TableComponent extends React.Component {
             renderInnerRCE={this.renderInnerRCE}
             onSelect={this.onSelect}
             theme={theme}
-            handleCopy={this.handleCopy}
             setRowRef={this.setRowRef}
             setEditorRef={this.setEditorRef}
             toolbarRef={this.toolbarRef}
@@ -477,6 +497,9 @@ class TableComponent extends React.Component {
             selectAll={isAllCellsSelected}
             tableHeight={this.tableRef.current?.offsetHeight}
             isEditingActive={this.state.isEditingActive}
+            onClear={this.table.clearCells}
+            setCellContent={this.setCellContent}
+            onPaste={this.onPaste}
           />
           <div className={styles.dragPreview} ref={this.dragPreview} />
         </div>