+++ packages/plugin-table/web/src/table-viewer.jsx
@@ -10,6 +10,7 @@ class TableViewer extends Component {
   constructor(props) {
     super(props);
     this.table = this.props.table || new TableDataUtil(props.componentData);
+    this.state = {};
   }
 
   componentDidMount() {
@@ -39,7 +40,6 @@ class TableViewer extends Component {
         {...props}
         table={this.table}
         tableRef={this.tableViewerRef}
-        isMobile={this.props.isMobile}
         colDragProps={this.props.colDragProps}
         onResize={this.props.onResize}
         onResizeStart={this.props.onResizeStart}
@@ -65,6 +65,9 @@ class TableViewer extends Component {
 
   setTableViewerRef = ref => {
     this.tableViewerRef = ref;
+    if (!this.state.isTableRefSet) {
+      this.setState({ isTableRefSet: true });
+    }
   };
 
   cellRenderer = props => {
@@ -100,7 +103,7 @@ class TableViewer extends Component {
   valueRenderer = cell => cell.component;
 
   render() {
-    const { onSelect, selected, handleCopy, isEditMode } = this.props;
+    const { onSelect, selected, isEditMode, setCellContent, onClear, onPaste } = this.props;
     const rowNum = this.table.getRowNum();
     const colNum = this.table.getColNum();
     this.grid = [...Array(rowNum).fill(0)].map((row, i) => this.createRow(i, colNum));
@@ -118,7 +121,10 @@ class TableViewer extends Component {
           cellRenderer={this.cellRenderer}
           rowRenderer={this.rowRenderer}
           sheetRenderer={this.sheetRenderer}
-          handleCopy={handleCopy}
+          onClear={onClear}
+          getCellContent={this.table.getCellContent}
+          setCellContent={setCellContent}
+          onPaste={onPaste}
         />
       </div>
     );
@@ -131,7 +137,6 @@ TableViewer.propTypes = {
   innerRCV: PropTypes.func,
   table: PropTypes.object,
   onSelect: PropTypes.func,
-  handleCopy: PropTypes.func,
   setRowRef: PropTypes.func,
   setEditorRef: PropTypes.func,
   toolbarRef: PropTypes.any,
@@ -153,6 +158,9 @@ TableViewer.propTypes = {
   selectAll: PropTypes.bool,
   tableHeight: PropTypes.number,
   isEditingActive: PropTypes.bool,
+  setCellContent: PropTypes.func,
+  onClear: PropTypes.func,
+  onPaste: PropTypes.func,
 };
 
 export default TableViewer;