+++ CHANGELOG.md
@@ -9,14 +9,67 @@
 > - :house: Internal
 > - :nail_care: Polish
 
-## [Unreleased]  
+## [Unreleased]
 <details>
   <summary>
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+
+### :rocket: New Feature
+- `common`
+  - [#2081](https://github.com/wix-incubator/rich-content/pull/2081) support for palette colors 3 & 4 (`disabledTextColor` & `textColorLow`)
+  - [#2079](https://github.com/wix-incubator/rich-content/pull/2079) experiments infra
+
+### :bug: Bug Fix
+- `unsupported-blocks`
+  - [##2078](https://github.com/wix/ricos/pull/#2078) fix unsupported-blocks container height issue 
+
 </details>
 <hr/>
+
+## 8.18.3 (Feb 2, 2021)
+### :bug: Bug Fix
+- `inner-modal`
+  - [#2076](https://github.com/wix-incubator/rich-content/pull/2076) vertical overflow issue
+## 8.18.1 (Feb 2, 2021)
+### :bug: Bug Fix
+- `ricos-editor`
+  - [#2073](https://github.com/wix-incubator/rich-content/pull/2073) isContentChanged fixed
+
+## 8.18.0 (Jan 31, 2021)
+### :rocket: New Feature
+- `ricos-common`
+  - [#2065](https://github.com/wix-incubator/rich-content/pull/2065) add textAlignment prop
+
+## 8.17.11 (Jan 27, 2021)
+### :bug: Bug Fix
+- `unsupported-blocks`
+  - [#2058](https://github.com/wix/ricos/pull/2058) bg color update
+
+
+## 8.17.10 (Jan 27, 2021)
+### :bug: Bug Fix
+- `link`
+  - [#2049](https://github.com/wix/ricos/pull/2049) fix url data changing rel/target when only one of them is present
+
+## 8.17.9 (Jan 26, 2021)
+### :bug: Bug Fix
+- `link`
+  - [#2038](https://github.com/wix/ricos/pull/2038) fix anchor's tag overriding url,add `href` value & remove `siteUrl` prop
+
+## 8.17.8 (Jan 25, 2021)
+### :rocket: New Feature
+- `editor`
+  - [#2034](https://github.com/wix-incubator/rich-content/pull/2034) maxTextLength prop added
+### :bug: Bug Fix
+- `table`
+  - [#2046](https://github.com/wix/ricos/pull/2046) cell selection background color style
+
+## 8.17.7 (Jan 25, 2021)
+### :bug: Bug Fix
+- `ricos-common`
+  - [#2041](https://github.com/wix/ricos/pull/2041) remove draft dependency in migration tool
 
 ## 8.17.6 (Jan 22, 2021)
 ### :rocket: New Feature
@@ -29,8 +82,6 @@
   - [#2026](https://github.com/wix/ricos/pull/2026) convert to TypeScript
 - `e2e`
   - [#2026](https://github.com/wix/ricos/pull/2026) convert test-env to TypeScript
-
-
 
 ## 8.17.5 (Jan 20, 2021)
 ### :house: Internal
@@ -248,7 +299,7 @@
 ### :bug: Bug Fix
 - `gallery`
   - [#1897](https://github.com/wix/ricos/pull/1897) fix gallery styles (college layout)
-  
+
 ## 8.9.2 (Dec 21, 2020)
 ### :bug: Bug Fix
 - `link`
@@ -320,7 +371,7 @@
 ### :bug: Bug Fix
 - `ricos-editor`
   - [#1850](https://github.com/wix/ricos/pull/1850) settings modal overlay appears above wix ad
-  
+
 ### :house: Internal
 - `test-env`
   - [#1848](https://github.com/wix/ricos/pull/1848) theme `customStyles` coverage